Add files via upload
I have 2 problems....I am trying to click the fabToCart button in the ProductActivity to put the shoe into the Cart.    Then there is a AddShoe (+) a textCounter and a subtractShoe when the user can select how many shoes to put in the cart.  I am stuck because I am not sure if the fabToCart code is right because when I run the Proust and click the fab button by a shoe, nothing happens.       Any assistance would helpful.    I am new at this.     Thanks in advance.